refactor(eval): require explicit engine runtimes
