+++ architecture/edge-baseline.json
@@ -29,8 +29,8 @@
   "legacy-runtime -> providers": 15,
   "legacy-runtime -> redteam": 6,
   "node -> core": 9,
-  "node -> legacy-contracts": 96,
-  "node -> legacy-runtime": 151,
+  "node -> legacy-contracts": 98,
+  "node -> legacy-runtime": 155,
   "node -> providers": 6,
   "node -> redteam": 10,
   "providers -> core": 1,
