+++ src/node/doEval.ts
@@ -12,6 +12,7 @@ import cliState from '../cliState';
 import { DEFAULT_MAX_CONCURRENCY } from '../constants';
 import { getEnvBool, getEnvFloat, getEnvInt, isCI } from '../envars';
 import { evaluate, PromptSuggestionsRejectedError } from '../evaluator';
+import { EvalRunError } from '../evaluator/errors';
 import {
   checkEmailStatusAndMaybeExit,
   EmailValidationError,
@@ -143,17 +144,8 @@ async function resolveReplayConfigs(
   return configs;
 }
 
-export class EvalRunError extends Error {
-  readonly exitCode: number;
-
-  constructor(message: string, exitCode: number = 1) {
-    super(message);
-    this.name = 'EvalRunError';
-    // POSIX exit codes are 1-255 (0 means success). Coerce silly values to the
-    // default rather than letting `exitCode = 0` silently mask a real failure.
-    this.exitCode = Number.isInteger(exitCode) && exitCode >= 1 && exitCode <= 255 ? exitCode : 1;
-  }
-}
+// Compatibility export for existing source consumers.
+export { EvalRunError } from '../evaluator/errors';
 
 function failEvalRun(
   message: string,
